requirements.txt
python-telegram-bot==20.7
playwright==1.40.0
python-dotenv==1.0.0
